fix(native-auth-ops): remove exceptions from logs in KDF restart function (#45)
* fix(native-auth-ops): remove find exceptions in log and wait for RPC

The `findExceptionsInLog` function called in the _restartKdf appears to be finding innocuous errors and re-throwing them. 

The one being caught on my machine is a socket connection error from an RPC, version, likely while KDF is starting or stopping.

Rethrowing exceptions from logs seems out-of-place in the restart kdf function.

* build: update coins commit and SDK example kdf_wrapper

* fix(log-censoring): add wallet_password to censored fields