refactor(macos): streamline KDF binary placement; update signing flow (#247)
* signing for all configurations in which it's needed

not only in "Release", runtime flags for signing in "Release",
"Release-production", etc.

* use thin_binary_to_archs for all Relase configurations

not only just "Release"

* change code signing approach, using code_sign_if_enabled function

* macos: move kdf binary from Resoures to Framework Helpers

to avoid issues with unsigned resources ... we put SIGNED kdf binary to
Contents/Frameworks/komodo_defi_framework.framework/Versions/Current/Helpers/kdf

In case of using Resources - it will result with unsigned binary (!)

TODO: May be this should be reworked, but we shouldn't store the binaries
in resources, to avoid issues.

* refactor(kdf_executable_finder): remove legacy macOS path for KDF executable

Eliminated the legacy path for the KDF executable in macOS to streamline the executable finding process. This change aligns with the recent restructuring of binary locations to avoid issues with unsigned resources.

* refactor(podspec): streamline macOS directory creation for KDF binary

Consolidated the creation of application support, frameworks, and helpers directories into a single command to simplify the build process.

---------

Co-authored-by: DeckerSU <deckersu@protonmail.com>